change to float

+++ database/migrations/2021_11_02_141418_create_timelogs_table.php
@@ -15,7 +15,7 @@ class CreateTimelogsTable extends Migration
     {
         Schema::create('timelogs', function (Blueprint $table) {
             $table->id();
-            $table->float('time');
+            $table->integer('time');
             $table->date('date');
             $table->string('comment');
             $table->timestamps();
